PHOTO_CAD/convert-orthophoto-to-dxf/pyproject.toml
[tool]
name = "convert-orthophoto-to-dxf"
version = "0.1.0"
description = "A project for converting orthophotos to DXF files with snapping functionality."
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/yourusername/convert-orthophoto-to-dxf"
repository = "https://github.com/yourusername/convert-orthophoto-to-dxf"
keywords = ["orthophoto", "DXF", "snapping", "vectorization"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.8"
ezdxf = "^0.16.0"
numpy = "^1.21.0"
opencv-python = "^4.5.3"
scikit-image = "^0.18.3"
pyyaml = "^5.4.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
black = "^21.7b0"
flake8 = "^3.9.2"

[tool.poetry.scripts]
convert-orthophoto-to-dxf = "convert_orthophoto_to_dxf_snapping:main"